feat: Allow specifying location by latitude/longitude
Allows the user to specify a custom location by passing a string
in "latitude;longitude" format to the `city` argument.

This provides a way to view the star sky from any arbitrary point
on Earth without needing a pre-defined city name.

The location parser supports several formats, including:
- Signed decimal degrees (e.g., "35.68;139.76")
- Degrees with direction characters (e.g., "N35.68;E139.76")

When a location is specified this way, the timezone is automatically
set to UTC.